+++ .storybook/main.js
@@ -63,4 +63,7 @@ module.exports = {
             propFilter: (prop) => (prop.parent ? !/node_modules/.test(prop.parent.fileName) : true),
         },
     },
+    reactOptions: {
+        fastRefresh: true,
+    },
 };